requirements.txt
# ============================================
# IoT Monitoring Dashboard - Requirements
# ICS-574: Big Data Analytics Course Project
# King Fahd University of Petroleum and Minerals
# ============================================

# ===========================================
# MESSAGING & STREAMING
# ===========================================
# MQTT client for IoT sensor communication
paho-mqtt>=1.6.1

# Kafka client for message broker integration
kafka-python>=2.0.2

# ===========================================
# STREAM PROCESSING (Apache Spark)
# ===========================================
# PySpark for real-time stream processing
pyspark>=3.5.0

# ===========================================
# DATABASE
# ===========================================
# PostgreSQL adapter for Python
psycopg2-binary>=2.9.9

# ===========================================
# DATA PROCESSING & ANALYSIS
# ===========================================
# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# ===========================================
# MACHINE LEARNING (Optional Smart Layer)
# ===========================================
# Scikit-learn for ML predictions
scikit-learn>=1.3.2

# ===========================================
# VISUALIZATION (Streamlit Dashboard)
# ===========================================
# Streamlit web application framework
streamlit>=1.28.0

# Interactive plotting library
plotly>=5.17.0

# System architecture diagrams
graphviz>=0.20.0
